requirements.txt
antiorm==1.2.1
backcall==0.1.0
colorama==0.3.9
debug==0.3.2
decorator==4.3.0
Django==2.1
django-bootstrap4==0.0.6
django-ckeditor==5.6.1
django-dbbackup==3.2.0
django-debug-toolbar==1.9.1
django-filter==2.0.0
django-forms-bootstrap==3.1.0
django-js-asset==1.1.0
django-markdown==0.8.4
django-menu==0.1.12
django-template-bootstrap==0.2.15
djangorestframework==3.8.2
fanstatic==1.0.0
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
js.bootstrap4==4.0.0
js.jquery==3.3.1
Markdown==2.6.11
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
Psycopg2Database==0.2.0
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
see==1.4.1
shutilwhich==1.1.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
wcwidth==0.1.7
WebOb==1.8.2
